move some of the mergeblob API over to unison-merge
