+++ unison-cli/src/Unison/Codebase/Editor/HandleInput/Merge2.hs
@@ -15,15 +15,9 @@ module Unison.Codebase.Editor.HandleInput.Merge2
   )
 where
 
-import Control.Lens (mapped)
 import Control.Monad.Reader (ask)
-import Data.Bifoldable (bifoldMap)
-import Data.List qualified as List
 import Data.Map.Strict qualified as Map
-import Data.Semialign (align, unzip, zipWith)
-import Data.Set qualified as Set
-import Data.Set.NonEmpty (NESet)
-import Data.Set.NonEmpty qualified as Set.NonEmpty
+import Data.Semialign (zipWith)
 import Data.Text qualified as Text
 import Data.Text.IO qualified as Text
 import Data.These (These (..))
@@ -49,7 +43,6 @@ import Unison.Cli.UpdateUtils
   ( getNamespaceDependentsOf3,
     hydrateDefns,
     loadNamespaceDefinitions,
-    renderDefnsForUnisonFile,
   )
 import Unison.Codebase (Codebase)
 import Unison.Codebase qualified as Codebase
@@ -67,7 +60,6 @@ import Unison.Codebase.ProjectPath qualified as PP
 import Unison.Codebase.SqliteCodebase.Branch.Cache (newBranchCache)
 import Unison.Codebase.SqliteCodebase.Conversions qualified as Conversions
 import Unison.Codebase.SqliteCodebase.Operations qualified as Operations
-import Unison.ConstructorReference (GConstructorReference (..))
 import Unison.ConstructorType (ConstructorType)
 import Unison.DataDeclaration (Decl)
 import Unison.DataDeclaration qualified as DataDeclaration
@@ -75,17 +67,13 @@ import Unison.Debug qualified as Debug
 import Unison.FileParsers qualified as FileParsers
 import Unison.Hash qualified as Hash
 import Unison.Merge qualified as Merge
-import Unison.Merge.DeclNameLookup (expectConstructorNames)
 import Unison.Merge.EitherWayI qualified as EitherWayI
 import Unison.Merge.Synhashed qualified as Synhashed
 import Unison.Merge.ThreeWay qualified as ThreeWay
-import Unison.Merge.TwoWay qualified as TwoWay
-import Unison.Merge.Unconflicts qualified as Unconflicts
 import Unison.Name (Name)
 import Unison.NameSegment (NameSegment)
 import Unison.NameSegment qualified as NameSegment
 import Unison.Names (Names (..))
-import Unison.Names qualified as Names
 import Unison.Parser.Ann (Ann)
 import Unison.Parsers qualified as Parsers
 import Unison.Prelude
@@ -97,7 +85,6 @@ import Unison.Project
     Semver (..),
     classifyProjectBranchName,
   )
-import Unison.Reference (Reference' (..), TermReference)
 import Unison.Reference qualified as Reference
 import Unison.Referent (Referent)
 import Unison.Referent qualified as Referent
@@ -115,14 +102,12 @@ import Unison.Typechecker qualified as Typechecker
 import Unison.Typechecker.TypeLookup (TypeLookup)
 import Unison.UnisonFile (TypecheckedUnisonFile, UnisonFile)
 import Unison.UnisonFile qualified as UnisonFile
-import Unison.Util.BiMultimap (BiMultimap)
 import Unison.Util.BiMultimap qualified as BiMultimap
 import Unison.Util.Conflicted (Conflicted)
 import Unison.Util.Defn (Defn)
-import Unison.Util.Defns (Defns (..), DefnsF, DefnsF2, DefnsF3, alignDefnsWith, defnsAreEmpty, zipDefnsWith, zipDefnsWith3)
+import Unison.Util.Defns (Defns (..), DefnsF, DefnsF2, DefnsF3, alignDefnsWith)
 import Unison.Util.Monoid qualified as Monoid
-import Unison.Util.Nametree (Nametree (..), flattenNametrees, unflattenNametree)
-import Unison.Util.Pretty (ColorText, Pretty)
+import Unison.Util.Nametree (Nametree (..), unflattenNametree)
 import Unison.Util.Pretty qualified as Pretty
 import Unison.Util.Relation (Relation)
 import Unison.Util.Relation qualified as Relation
@@ -253,7 +238,7 @@ doMerge info = do
 
       libdeps3 <- Cli.runTransaction (loadLibdeps branches)
 
-      let blob0 = makeMergeblob0 nametrees3 libdeps3
+      let blob0 = Merge.makeMergeblob0 nametrees3 libdeps3
 
       -- Hydrate
       hydratedDefns ::
@@ -275,7 +260,7 @@ doMerge info = do
             )
 
       blob1 <-
-        makeMergeblob1 blob0 hydratedDefns & onLeft \case
+        Merge.makeMergeblob1 blob0 hydratedDefns & onLeft \case
           Merge.Alice reason -> done (Output.IncoherentDeclDuringMerge mergeTarget reason)
           Merge.Bob reason -> done (Output.IncoherentDeclDuringMerge mergeSource reason)
 
@@ -284,13 +269,13 @@ doMerge info = do
       liftIO (debugFunctions.debugCombinedDiff blob1.diff)
 
       blob2 <-
-        makeMergeblob2 blob1 & onLeft \err ->
+        Merge.makeMergeblob2 blob1 & onLeft \err ->
           done case err of
-            Mergeblob2Error'ConflictedAlias defn0 ->
+            Merge.Mergeblob2Error'ConflictedAlias defn0 ->
               case defn0 of
                 Merge.Alice defn -> Output.MergeConflictedAliases mergeTarget defn
                 Merge.Bob defn -> Output.MergeConflictedAliases mergeSource defn
-            Mergeblob2Error'ConflictedBuiltin defn -> Output.MergeConflictInvolvingBuiltin defn
+            Merge.Mergeblob2Error'ConflictedBuiltin defn -> Output.MergeConflictInvolvingBuiltin defn
 
       liftIO (debugFunctions.debugPartitionedDiff blob2.conflicts blob2.unconflicts)
 
@@ -309,7 +294,7 @@ doMerge info = do
             blob2.hasConflicts
 
       let blob3 =
-            makeMergeblob3
+            Merge.makeMergeblob3
               blob2
               dependents0
               (Branch.toNames mergedLibdeps)
@@ -372,248 +357,12 @@ doMerge info = do
 
   Cli.respond finalOutput
 
-data Mergeblob0 libdep = Mergeblob0
-  { defns :: Merge.ThreeWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name)),
-    libdeps :: Merge.ThreeWay (Map NameSegment libdep),
-    nametrees :: Merge.ThreeWay (Nametree (DefnsF (Map NameSegment) Referent TypeReference))
-  }
-
-makeMergeblob0 ::
-  Merge.ThreeWay (Nametree (DefnsF (Map NameSegment) Referent TypeReference)) ->
-  Merge.ThreeWay (Map NameSegment libdep) ->
-  Mergeblob0 libdep
-makeMergeblob0 nametrees libdeps =
-  Mergeblob0
-    { defns = flattenNametrees <$> nametrees,
-      libdeps,
-      nametrees
-    }
-
-data Mergeblob1 libdep = Mergeblob1
-  { conflicts :: Merge.TwoWay (DefnsF (Map Name) TermReference TypeReference),
-    declNameLookups :: Merge.TwoWay Merge.DeclNameLookup,
-    defns :: Merge.ThreeWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name)),
-    diff :: DefnsF2 (Map Name) Merge.CombinedDiffOp Referent TypeReference,
-    diffs :: Merge.TwoWay (DefnsF3 (Map Name) Merge.DiffOp Merge.Synhashed Referent TypeReference),
-    hydratedDefns ::
-      Merge.ThreeWay
-        ( DefnsF
-            (Map Name)
-            (TermReferenceId, (Term Symbol Ann, Type Symbol Ann))
-            (TypeReferenceId, Decl Symbol Ann)
-        ),
-    lcaDeclNameLookup :: Merge.PartialDeclNameLookup,
-    libdeps :: Map NameSegment libdep,
-    libdepsDiff :: Map NameSegment (Merge.LibdepDiffOp libdep),
-    unconflicts :: DefnsF Merge.Unconflicts Referent TypeReference
-  }
-
-makeMergeblob1 ::
-  forall libdep.
-  (Eq libdep) =>
-  Mergeblob0 libdep ->
-  Merge.ThreeWay
-    ( DefnsF
-        (Map Name)
-        (TermReferenceId, (Term Symbol Ann, Type Symbol Ann))
-        (TypeReferenceId, Decl Symbol Ann)
-    ) ->
-  Either (Merge.EitherWay Merge.IncoherentDeclReason) (Mergeblob1 libdep)
-makeMergeblob1 blob hydratedDefns = do
-  -- Make one big constructor count lookup for all type decls
-  let numConstructors =
-        Map.empty
-          & f (Map.elems hydratedDefns.alice.types)
-          & f (Map.elems hydratedDefns.bob.types)
-          & f (Map.elems hydratedDefns.lca.types)
-        where
-          f :: [(TypeReferenceId, Decl Symbol Ann)] -> Map TypeReferenceId Int -> Map TypeReferenceId Int
-          f types acc =
-            List.foldl'
-              ( \acc (ref, decl) ->
-                  Map.insert ref (DataDeclaration.constructorCount (DataDeclaration.asDataDecl decl)) acc
-              )
-              acc
-              types
-
-  -- Make Alice/Bob decl name lookups, which can fail if either have an incoherent decl
-  declNameLookups <- do
-    alice <- Merge.checkDeclCoherency blob.nametrees.alice numConstructors & mapLeft Merge.Alice
-    bob <- Merge.checkDeclCoherency blob.nametrees.bob numConstructors & mapLeft Merge.Bob
-    pure Merge.TwoWay {alice, bob}
-
-  -- Make LCA decl name lookup
-  let lcaDeclNameLookup =
-        Merge.lenientCheckDeclCoherency blob.nametrees.lca numConstructors
-
-  -- Diff LCA->Alice and LCA->Bob
-  let diffs =
-        Merge.nameBasedNamespaceDiff
-          declNameLookups
-          lcaDeclNameLookup
-          blob.defns
-          Defns
-            { terms =
-                foldMap
-                  (List.foldl' (\acc (ref, (term, _)) -> Map.insert ref term acc) Map.empty . Map.elems . (.terms))
-                  hydratedDefns,
-              types =
-                foldMap
-                  (List.foldl' (\acc (ref, typ) -> Map.insert ref typ acc) Map.empty . Map.elems . (.types))
-                  hydratedDefns
-            }
-
-  -- Combine the LCA->Alice and LCA->Bob diffs together
-  let diff =
-        Merge.combineDiffs diffs
-
-  -- Partition the combined diff into the conflicted things and the unconflicted things
-  let (conflicts, unconflicts) =
-        Merge.partitionCombinedDiffs (ThreeWay.forgetLca blob.defns) declNameLookups diff
-
-  -- Diff and merge libdeps
-  let libdepsDiff :: Map NameSegment (Merge.LibdepDiffOp libdep)
-      libdepsDiff =
-        Merge.diffLibdeps blob.libdeps
-
-  let libdeps :: Map NameSegment libdep
-      libdeps =
-        Merge.applyLibdepsDiff Merge.getTwoFreshLibdepNames blob.libdeps libdepsDiff
-
-  pure
-    Mergeblob1
-      { conflicts,
-        declNameLookups,
-        defns = blob.defns,
-        diff,
-        diffs,
-        hydratedDefns,
-        lcaDeclNameLookup,
-        libdeps,
-        libdepsDiff,
-        unconflicts
-      }
-
-data Mergeblob2 libdep = Mergeblob2
-  { conflicts :: Merge.TwoWay (DefnsF (Map Name) TermReferenceId TypeReferenceId),
-    conflictsIds :: Merge.TwoWay (DefnsF Set TermReferenceId TypeReferenceId),
-    conflictsNames :: Merge.TwoWay (DefnsF Set Name Name),
-    coreDependencies :: Merge.TwoWay (DefnsF Set TermReference TypeReference),
-    declNameLookups :: Merge.TwoWay Merge.DeclNameLookup,
-    defns :: Merge.ThreeWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name)),
-    hasConflicts :: Bool,
-    hydratedDefns ::
-      Merge.TwoWay
-        ( DefnsF
-            (Map Name)
-            (TermReferenceId, (Term Symbol Ann, Type Symbol Ann))
-            (TypeReferenceId, Decl Symbol Ann)
-        ),
-    lcaDeclNameLookup :: Merge.PartialDeclNameLookup,
-    libdeps :: Map NameSegment libdep,
-    soloUpdatesAndDeletes :: Merge.TwoWay (DefnsF Set Name Name),
-    unconflicts :: DefnsF Merge.Unconflicts Referent TypeReference
-  }
-
-data Mergeblob2Error
-  = Mergeblob2Error'ConflictedAlias (Merge.EitherWay (Defn (Name, Name) (Name, Name)))
-  | Mergeblob2Error'ConflictedBuiltin (Defn Name Name)
-
-makeMergeblob2 :: Mergeblob1 libdep -> Either Mergeblob2Error (Mergeblob2 libdep)
-makeMergeblob2 blob = do
-  -- Bail early if it looks like we can't proceed with the merge, because Alice or Bob has one or more conflicted alias
-  for_ ((,) <$> Merge.TwoWay Merge.Alice Merge.Bob <*> blob.diffs) \(who, diff) ->
-    whenJust (Merge.findConflictedAlias blob.defns.lca diff) $
-      Left . Mergeblob2Error'ConflictedAlias . who
-
-  conflicts <- Merge.narrowConflictsToNonBuiltins blob.conflicts & mapLeft Mergeblob2Error'ConflictedBuiltin
-  let conflictsIds = bimap (Set.fromList . Map.elems) (Set.fromList . Map.elems) <$> conflicts
-  let conflictsNames = bimap Map.keysSet Map.keysSet <$> conflicts
-
-  let soloUpdatesAndDeletes = Unconflicts.soloUpdatesAndDeletes blob.unconflicts
-  let coreDependencies = identifyCoreDependencies (ThreeWay.forgetLca blob.defns) conflictsIds soloUpdatesAndDeletes
-
-  pure
-    Mergeblob2
-      { conflicts,
-        conflictsIds,
-        conflictsNames,
-        coreDependencies,
-        declNameLookups = blob.declNameLookups,
-        defns = blob.defns,
-        -- Eh, they'd either both be null, or neither, but just check both maps anyway
-        hasConflicts = not (defnsAreEmpty conflicts.alice) || not (defnsAreEmpty conflicts.bob),
-        hydratedDefns = ThreeWay.forgetLca blob.hydratedDefns,
-        lcaDeclNameLookup = blob.lcaDeclNameLookup,
-        libdeps = blob.libdeps,
-        soloUpdatesAndDeletes,
-        unconflicts = blob.unconflicts
-      }
-
-data Mergeblob3 = Mergeblob3
-  { libdeps :: Names,
-    stageOne :: DefnsF (Map Name) Referent TypeReference,
-    unparsedFile :: Pretty ColorText
-  }
-
-makeMergeblob3 ::
-  Mergeblob2 libdep ->
-  Merge.TwoWay (DefnsF Set TermReferenceId TypeReferenceId) ->
-  Names ->
-  Merge.TwoWay Text ->
-  Mergeblob3
-makeMergeblob3 blob dependents0 libdeps authors =
-  -- Identify the unconflicted dependents we need to pull into the Unison file (either first for typechecking, if
-  -- there aren't conflicts, or else for manual conflict resolution without a typechecking step, if there are)
-  let dependents =
-        filterDependents
-          blob.conflictsNames
-          blob.soloUpdatesAndDeletes
-          ( let f :: Set TermReferenceId -> Referent -> NESet Name -> Set Name
-                f deps defn0 names
-                  | Just defn <- Referent.toTermReferenceId defn0,
-                    Set.member defn deps =
-                      Set.NonEmpty.toSet names
-                  | otherwise = Set.empty
-                g :: Set TypeReferenceId -> TypeReference -> NESet Name -> Set Name
-                g deps defn0 names
-                  | ReferenceDerived defn <- defn0,
-                    Set.member defn deps =
-                      Set.NonEmpty.toSet names
-                  | otherwise = Set.empty
-             in zipDefnsWith
-                  (\defns deps -> Map.foldMapWithKey (f deps) (BiMultimap.domain defns))
-                  (\defns deps -> Map.foldMapWithKey (g deps) (BiMultimap.domain defns))
-                  <$> ThreeWay.forgetLca blob.defns
-                  <*> dependents0
-          )
-
-      (renderedConflicts, renderedDependents) =
-        renderConflictsAndDependents
-          blob.declNameLookups
-          blob.hydratedDefns
-          blob.conflictsNames
-          dependents
-          (defnsToNames <$> ThreeWay.forgetLca blob.defns)
-          libdeps
-   in Mergeblob3
-        { libdeps,
-          stageOne =
-            makeStageOne
-              blob.declNameLookups
-              blob.conflictsNames
-              blob.unconflicts
-              dependents
-              (bimap BiMultimap.range BiMultimap.range blob.defns.lca),
-          unparsedFile = makePrettyUnisonFile authors renderedConflicts renderedDependents
-        }
-
 data Mergeblob4 = Mergeblob4
   { dependencies :: Set Reference,
     file :: UnisonFile Symbol Ann
   }
 
-makeMergeblob4 :: Mergeblob3 -> UniqueName -> Either (Parser.Err Symbol) Mergeblob4
+makeMergeblob4 :: Merge.Mergeblob3 -> UniqueName -> Either (Parser.Err Symbol) Mergeblob4
 makeMergeblob4 blob uniqueName = do
   let stageOneNames =
         Names (Relation.fromMap blob.stageOne.terms) (Relation.fromMap blob.stageOne.types) <> blob.libdeps
@@ -649,41 +398,6 @@ makeMergeblob5 blob typeLookup =
    in case runIdentity (Result.runResultT (FileParsers.synthesizeFile typecheckingEnv blob.file)) of
         (Nothing, notes) -> Left notes
         (Just file, _) -> Right Mergeblob5 {file}
-
-renderConflictsAndDependents ::
-  Merge.TwoWay Merge.DeclNameLookup ->
-  Merge.TwoWay (DefnsF (Map Name) (TermReferenceId, (Term Symbol Ann, Type Symbol Ann)) (TypeReferenceId, Decl Symbol Ann)) ->
-  Merge.TwoWay (DefnsF Set Name Name) ->
-  Merge.TwoWay (DefnsF Set Name Name) ->
-  Merge.TwoWay Names ->
-  Names ->
-  ( Merge.TwoWay (DefnsF (Map Name) (Pretty ColorText) (Pretty ColorText)),
-    Merge.TwoWay (DefnsF (Map Name) (Pretty ColorText) (Pretty ColorText))
-  )
-renderConflictsAndDependents declNameLookups hydratedDefns conflicts dependents names libdepsNames =
-  unzip $
-    ( \declNameLookup (conflicts, dependents) ppe ->
-        let render = renderDefnsForUnisonFile declNameLookup ppe . over (#terms . mapped) snd
-         in (render conflicts, render dependents)
-    )
-      <$> declNameLookups
-      <*> hydratedConflictsAndDependents
-      <*> Merge.makePrettyPrintEnvs names libdepsNames
-  where
-    hydratedConflictsAndDependents ::
-      Merge.TwoWay
-        ( DefnsF (Map Name) (TermReferenceId, (Term Symbol Ann, Type Symbol Ann)) (TypeReferenceId, Decl Symbol Ann),
-          DefnsF (Map Name) (TermReferenceId, (Term Symbol Ann, Type Symbol Ann)) (TypeReferenceId, Decl Symbol Ann)
-        )
-    hydratedConflictsAndDependents =
-      ( \as bs cs ->
-          ( zipDefnsWith Map.restrictKeys Map.restrictKeys as bs,
-            zipDefnsWith Map.restrictKeys Map.restrictKeys as cs
-          )
-      )
-        <$> hydratedDefns
-        <*> conflicts
-        <*> dependents
 
 doMergeLocalBranch :: Merge.TwoWay (ProjectAndBranch Project ProjectBranch) -> Cli ()
 doMergeLocalBranch branches = do
@@ -750,103 +464,7 @@ hasDefnsInLib branch = do
   pure (not (Map.null libdeps.terms) || not (Map.null libdeps.types))
 
 ------------------------------------------------------------------------------------------------------------------------
--- Creating Unison files
-
-makePrettyUnisonFile ::
-  Merge.TwoWay Text ->
-  Merge.TwoWay (DefnsF (Map Name) (Pretty ColorText) (Pretty ColorText)) ->
-  Merge.TwoWay (DefnsF (Map Name) (Pretty ColorText) (Pretty ColorText)) ->
-  Pretty ColorText
-makePrettyUnisonFile authors conflicts dependents =
-  fold
-    [ conflicts
-        -- Merge the two maps together into one, remembering who authored what
-        & TwoWay.twoWay (zipDefnsWith align align)
-        -- Sort alphabetically
-        & inAlphabeticalOrder
-        -- Render each conflict, types then terms (even though a type can conflict with a term, in which case they
-        -- would not be adjacent in the file), with an author comment above each conflicted thing
-        & ( let f =
-                  foldMap \case
-                    This x -> alice x
-                    That y -> bob y
-                    These x y -> alice x <> bob y
-                  where
-                    alice = prettyBinding (Just (Pretty.text authors.alice))
-                    bob = prettyBinding (Just (Pretty.text authors.bob))
-             in bifoldMap f f
-          ),
-      -- Show message that delineates where conflicts end and dependents begin only when there are both conflicts and
-      -- dependents
-      let thereAre defns = TwoWay.or (not . defnsAreEmpty <$> defns)
-       in if thereAre conflicts && thereAre dependents
-            then
-              fold
-                [ "-- The definitions below are not conflicted, but they each depend on one or more\n",
-                  "-- conflicted definitions above.\n\n"
-                ]
-            else mempty,
-      dependents
-        -- Merge dependents together into one map (they are disjoint)
-        & TwoWay.twoWay (zipDefnsWith Map.union Map.union)
-        -- Sort alphabetically
-        & inAlphabeticalOrder
-        -- Render each dependent, types then terms, without bothering to comment attribution
-        & (let f = foldMap (prettyBinding Nothing) in bifoldMap f f)
-    ]
-  where
-    prettyBinding maybeComment binding =
-      fold
-        [ case maybeComment of
-            Nothing -> mempty
-            Just comment -> "-- " <> comment <> "\n",
-          binding,
-          "\n",
-          "\n"
-        ]
-
-    inAlphabeticalOrder :: DefnsF (Map Name) a b -> DefnsF [] a b
-    inAlphabeticalOrder =
-      bimap f f
-      where
-        f = map snd . List.sortOn (Name.toText . fst) . Map.toList
-
-------------------------------------------------------------------------------------------------------------------------
 --
-
--- Given just named term/type reference ids, fill out all names that occupy the term and type namespaces. This is simply
--- the given names plus all of the types' constructors.
---
--- For example, if the input is
---
---   declNameLookup = {
---     "Maybe" => ["Maybe.Nothing", "Maybe.Just"]
---   }
---   defns = {
---     terms = { "foo" => #foo }
---     types = { "Maybe" => #Maybe }
---   }
---
--- then the output is
---
---   defns = {
---     terms = { "foo", "Maybe.Nothing", "Maybe.Just" }
---     types = { "Maybe" }
---   }
-refIdsToNames :: Merge.DeclNameLookup -> DefnsF Set Name Name -> DefnsF Set Name Name
-refIdsToNames declNameLookup =
-  bifoldMap goTerms goTypes
-  where
-    goTerms :: Set Name -> DefnsF Set Name Name
-    goTerms terms =
-      Defns {terms, types = Set.empty}
-
-    goTypes :: Set Name -> DefnsF Set Name Name
-    goTypes types =
-      Defns
-        { terms = foldMap (Set.fromList . expectConstructorNames declNameLookup) types,
-          types
-        }
 
 defnsAndLibdepsToBranch0 ::
   Codebase IO v a ->
@@ -892,115 +510,6 @@ nametreeToBranch0 nametree =
     rel2star :: Relation ref name -> Star2 ref name metadata
     rel2star rel =
       Star2.Star2 {fact = Relation.dom rel, d1 = rel, d2 = Relation.empty}
-
-identifyCoreDependencies ::
-  Merge.TwoWay (Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name)) ->
-  Merge.TwoWay (DefnsF Set TermReferenceId TypeReferenceId) ->
-  Merge.TwoWay (DefnsF Set Name Name) ->
-  Merge.TwoWay (DefnsF Set TermReference TypeReference)
-identifyCoreDependencies defns conflicts soloUpdatesAndDeletes = do
-  fold
-    [ -- One source of dependencies: Alice's versions of Bob's unconflicted deletes and updates, and vice-versa.
-      --
-      -- This is name-based: if Bob updates the *name* "foo", then we go find the thing that Alice calls "foo" (if
-      -- anything), no matter what its hash is.
-      defnsReferences
-        <$> ( zipDefnsWith BiMultimap.restrictRan BiMultimap.restrictRan
-                <$> TwoWay.swap soloUpdatesAndDeletes
-                <*> defns
-            ),
-      -- The other source of dependencies: Alice's own conflicted things, and ditto for Bob.
-      --
-      -- An example: suppose Alice has foo#alice and Bob has foo#bob, so foo is conflicted. Furthermore, suppose
-      -- Alice has bar#bar that depends on foo#alice.
-      --
-      -- We want Alice's #alice to be considered a dependency, so that when we go off and find dependents of these
-      -- dependencies to put in the scratch file for type checking and propagation, we find bar#bar.
-      --
-      -- Note that this is necessary even if bar#bar is unconflicted! We don't want bar#bar to be put directly
-      -- into the namespace / parsing context for the conflicted merge, because it has an unnamed reference on
-      -- foo#alice. It rather ought to be in the scratchfile alongside the conflicted foo#alice and foo#bob, so
-      -- that when that conflict is resolved, it will propagate to bar.
-      bimap (Set.map Reference.DerivedId) (Set.map Reference.DerivedId) <$> conflicts
-    ]
-
-filterDependents ::
-  (Ord name) =>
-  Merge.TwoWay (DefnsF Set name name) ->
-  Merge.TwoWay (DefnsF Set name name) ->
-  Merge.TwoWay (DefnsF Set name name) ->
-  Merge.TwoWay (DefnsF Set name name)
-filterDependents conflicts soloUpdatesAndDeletes dependents0 =
-  -- There is some subset of Alice's dependents (and ditto for Bob of course) that we don't ultimately want/need to put
-  -- into the scratch file: those for which any of the following are true:
-  --
-  --   1. It is Alice-conflicted (since we only want to return *unconflicted* things).
-  --   2. It was deleted by Bob.
-  --   3. It was updated by Bob and not updated by Alice.
-  let dependents1 =
-        zipDefnsWith Set.difference Set.difference
-          <$> dependents0
-          <*> (conflicts <> TwoWay.swap soloUpdatesAndDeletes)
-
-      -- Of the remaining dependents, it's still possible that the maps are not disjoint. But whenever the same name key
-      -- exists in Alice's and Bob's dependents, the value will either be equal (by Unison hash)...
-      --
-      --   { alice = { terms = {"foo" => #alice} } }
-      --   { bob   = { terms = {"foo" => #alice} } }
-      --
-      -- ...or synhash-equal (i.e. the term or type received different auto-propagated updates)...
-      --
-      --   { alice = { terms = {"foo" => #alice} } }
-      --   { bob   = { terms = {"foo" => #bob}   } }
-      --
-      -- So, we can arbitrarily keep Alice's, because they will render the same.
-      --
-      --   { alice = { terms = {"foo" => #alice} } }
-      --   { bob   = { terms = {}                } }
-      dependents2 =
-        dependents1 & over #bob \bob ->
-          zipDefnsWith Set.difference Set.difference bob dependents1.alice
-   in dependents2
-
-makeStageOne ::
-  Merge.TwoWay Merge.DeclNameLookup ->
-  Merge.TwoWay (DefnsF Set Name Name) ->
-  DefnsF Merge.Unconflicts term typ ->
-  Merge.TwoWay (DefnsF Set Name Name) ->
-  DefnsF (Map Name) term typ ->
-  DefnsF (Map Name) term typ
-makeStageOne declNameLookups conflicts unconflicts dependents =
-  zipDefnsWith3 makeStageOneV makeStageOneV unconflicts (f conflicts <> f dependents)
-  where
-    f :: Merge.TwoWay (DefnsF Set Name Name) -> DefnsF Set Name Name
-    f defns =
-      fold (refIdsToNames <$> declNameLookups <*> defns)
-
-makeStageOneV :: Merge.Unconflicts v -> Set Name -> Map Name v -> Map Name v
-makeStageOneV unconflicts namesToDelete =
-  (`Map.withoutKeys` namesToDelete) . Unconflicts.apply unconflicts
-
-defnsReferences ::
-  Defns (BiMultimap Referent name) (BiMultimap TypeReference name) ->
-  DefnsF Set TermReference TypeReference
-defnsReferences defns =
-  List.foldl' f Defns {terms = Set.empty, types = BiMultimap.dom defns.types} (Set.toList (BiMultimap.dom defns.terms))
-  where
-    f :: DefnsF Set TermReference TypeReference -> Referent -> DefnsF Set TermReference TypeReference
-    f acc = \case
-      Referent.Con (ConstructorReference ref _) _ ->
-        let !types = Set.insert ref acc.types
-         in Defns {terms = acc.terms, types}
-      Referent.Ref ref ->
-        let !terms = Set.insert ref acc.terms
-         in Defns {terms, types = acc.types}
-
-defnsToNames :: Defns (BiMultimap Referent Name) (BiMultimap TypeReference Name) -> Names
-defnsToNames defns =
-  Names.Names
-    { terms = Relation.fromMap (BiMultimap.range defns.terms),
-      types = Relation.fromMap (BiMultimap.range defns.types)
-    }
 
 findTemporaryBranchName :: ProjectId -> MergeSourceAndTarget -> Transaction ProjectBranchName
 findTemporaryBranchName projectId mergeSourceAndTarget = do
