+++ unison-merge/package.yaml
@@ -31,6 +31,7 @@ dependencies:
   - unison-hash
   - unison-parser-typechecker
   - unison-prelude
+  - unison-pretty-printer
   - unison-sqlite
   - unison-syntax
   - unison-util-cache
