+++ unison-merge/src/Unison/Merge.hs
@@ -1,5 +1,16 @@
 module Unison.Merge
-  ( -- * Decl coherency checks
+  ( Mergeblob0 (..),
+    makeMergeblob0,
+    Mergeblob1 (..),
+    makeMergeblob1,
+    Mergeblob2 (..),
+    Mergeblob2Error (..),
+    makeMergeblob2,
+    Mergeblob3 (..),
+    makeMergeblob3,
+
+    -- * Decl coherency checks
+
     DeclNameLookup (..),
     PartialDeclNameLookup (..),
     IncoherentDeclReason (..),
@@ -9,38 +20,48 @@ module Unison.Merge
     checkAllDeclCoherency,
 
     -- * 3-way namespace diff
-    DiffOp (..),
-    nameBasedNamespaceDiff,
+
+    -- DiffOp (..),
+    -- nameBasedNamespaceDiff,
 
     -- * Finding conflicted aliases
-    findConflictedAlias,
+
+    -- findConflictedAlias,
 
     -- * Combining namespace diffs
-    CombinedDiffOp (..),
-    combineDiffs,
+
+    -- CombinedDiffOp (..),
+    -- combineDiffs,
 
     -- * Partitioning combined namespace diffs
-    Unconflicts (..),
-    partitionCombinedDiffs,
-    narrowConflictsToNonBuiltins,
+
+    -- Unconflicts (..),
+    -- partitionCombinedDiffs,
+    -- narrowConflictsToNonBuiltins,
 
     -- * Merging libdeps
-    LibdepDiffOp (..),
-    diffLibdeps,
-    applyLibdepsDiff,
-    getTwoFreshLibdepNames,
+
+    -- LibdepDiffOp (..),
+    -- diffLibdeps,
+    -- applyLibdepsDiff,
+    -- getTwoFreshLibdepNames,
 
     -- * Making a pretty-print environment
-    makePrettyPrintEnvs,
 
-    -- * Utility types
+    -- makePrettyPrintEnvs,
+
+    -- * Types
+    CombinedDiffOp(..),
+    DiffOp(..),
     EitherWay (..),
+    EitherWayI (..),
+    LibdepDiffOp(..),
+    Synhashed (..),
     ThreeWay (..),
     TwoOrThreeWay (..),
-    EitherWayI (..),
-    Synhashed (..),
     TwoWay (..),
     TwoWayI (..),
+    Unconflicts(..),
     Updated (..),
   )
 where
@@ -60,6 +81,10 @@ import Unison.Merge.EitherWay (EitherWay (..))
 import Unison.Merge.EitherWayI (EitherWayI (..))
 import Unison.Merge.FindConflictedAlias (findConflictedAlias)
 import Unison.Merge.Libdeps (LibdepDiffOp (..), applyLibdepsDiff, diffLibdeps, getTwoFreshLibdepNames)
+import Unison.Merge.Mergeblob0 (Mergeblob0 (..), makeMergeblob0)
+import Unison.Merge.Mergeblob1 (Mergeblob1 (..), makeMergeblob1)
+import Unison.Merge.Mergeblob2 (Mergeblob2 (..), Mergeblob2Error (..), makeMergeblob2)
+import Unison.Merge.Mergeblob3 (Mergeblob3 (..), makeMergeblob3)
 import Unison.Merge.PartialDeclNameLookup (PartialDeclNameLookup (..))
 import Unison.Merge.PartitionCombinedDiffs (narrowConflictsToNonBuiltins, partitionCombinedDiffs)
 import Unison.Merge.PrettyPrintEnv (makePrettyPrintEnvs)
